Account for class size upon object initialization
Added accounting for object size (based on its field count) upon
initialization. This prevents an adversary from allocating many
'large' objects with many fields to cause out of memory exceptions.

Before a NEW opcode, the recorder is given a string for the className.
The recorder then checks a global mapping of classNames to their
field counts. If such a mapping doesn't exist for a className, then
the recorder must load / instrument the class. It currently does this
by using the loadClassForAnalysis method in MyClassLoader. Currently,
this method does more work than it needs to, but it appears to have
reasonable behavior. The file resources/HelloWorld.java has been updated
with a few lines to test its functionality.